gameplay is now in the Ingame comp
but its still not connected to the ui. Additionally its now impossible to enter /play without matchmaking
