+++ truco-front/src/components/examples/GameLogicTest.tsx
@@ -1,15 +1,11 @@
-import { useEffect, useRef, useState } from "react";
+import { useEffect, useState } from "react";
 import GameManager from '../../gameLogic/GameManager';
 import { Card } from '../../gameLogic/Cards/Card';
 import { GameActionMessage } from '../../gameLogic/type/GameActionMessage';
 import { GameAction } from '../../gameLogic/type/GameAction';
-import Timer from "../matchmaking/matchmakingTimer";
-import { usePusherListeners } from "../../hooks/usePusherListeners";
 
 function GameLogicTest() {
     const [isSearching, setIsSearching] = useState<boolean>(false);
-    const isSearchingRef = useRef<boolean>(false);
-
     const [opponentName, setOpponentName] = useState<string>("");
     const [isMyTurn, setIsMyTurn] = useState<boolean>(false);
     const [actions, setActions] = useState<GameAction[]>([])
@@ -17,23 +13,66 @@ function GameLogicTest() {
     const gameManager = GameManager.getInstance()
 
 
-    usePusherListeners(gameManager, setOpponentName, setCards, setActions, setIsMyTurn)
+    useEffect(() => {
+        const username = "test-name" + Math.floor(Math.random() * 10000) // TODO - replace with the real username once we have users
+        gameManager.initPusher(username)
+        gameManager.events.addMatchFoundListener((opponentName: string) => {
+            setOpponentName(opponentName)
+        })
+        gameManager.events.addOnGetCardsListener((cards: Card[]) => {
+            setCards(cards)
+        })
+        gameManager.events.addOnMyTurnStartListener(() => {
+            console.log("my turn start")
+            const possibleActions = gameManager.getPossibleActions()
+            const newActions = []
+            for (let action of possibleActions) {
+                if (action[1]) {
+                    newActions.push(action[0])
+                }
+            }
+            setActions(newActions)
+            setIsMyTurn(true)
+        })
+        gameManager.events.addOnMyTurnEndListener(() => {
+            console.log("my turn end")
+            setIsMyTurn(false)
+            setActions([])
+        })
+
+        gameManager.events.addOnCardPlayedListener((iCalled: boolean, card: Card) => {
+            if (!iCalled) {
+                console.log("opponent played card: " + card.number + " " + card.suit)
+                return;
+            }
+            console.log("i played card: " + card.number + " " + card.suit)
+            const cardIndex = cards.findIndex((c) => c.number === card.number && c.suit === card.suit)
+            cards.splice(cardIndex, 1)
+        })
+
+        gameManager.events.addOnGameEndListener((IWon: boolean) => {
+            console.log("game end")
+            console.log(IWon)
+            setIsMyTurn(false)
+            setActions([])
+            setOpponentName("")
+            setCards([])
+            setIsSearching(false)
+        })
+        gameManager.events.addOnPointsUpdateListener((myPoints: number, opponentPoints: number) => {
+            console.log("points update")
+            console.log(myPoints)
+            console.log(opponentPoints)
+        })
+    }, [])
 
     const toggleMatchmaking = () => {
-        if (isSearchingRef.current) {
-            setIsSearching(false)
-            isSearchingRef.current = false
+        setIsSearching(!isSearching)
+        if (isSearching) {
             gameManager.leaveMatchmaking()
-            return
-        }
-
-        setIsSearching(true)
-        isSearchingRef.current = true
-
-        setTimeout(() => {
-            if (!isSearchingRef.current) return
+        } else {
             gameManager.joinMatchmaking()
-        }, 4000)
+        }
     }
 
     const playAction = (actionName: GameAction) => {
@@ -49,7 +88,6 @@ function GameLogicTest() {
     return (
         <div>
             <button className={buttonStyle} onClick={toggleMatchmaking}>"{isSearching ? "Stop" : "Play"}"</button>
-            {isSearching && <Timer />}
             {opponentName && <div>Match found! Opponent: {opponentName}</div>}
             {isMyTurn ? actions.map((action, index) => {
                 return <button className={buttonStyle} key={index} onClick={() => playAction(action)}>{action}</button>
