+++ truco-front/src/components/inGame/GameBoard.tsx
@@ -15,6 +15,7 @@ const myCards = [
 ]
 
 const GameBoard = () => {
+
     return (
         <div className="w-[715px] h-[715px] m-auto flex relative justify-evenly items-center rounded-2xl bg-[url('assets/mesa.jpg')] shadow-card">
             <TrucoActions />
