+++ truco-front/src/hooks/usePusherListeners.ts
@@ -9,7 +9,7 @@ export const usePusherListeners = (
     setOpponentName: (name: string) => void,
     setCards: Function,
     setActions: (actions: GameAction[]) => void,
-    setIsMyTurn: (isMyTurn: boolean) => void
+    setIsMyTurn: (isMyTurn: boolean) => void,
 ) => {
 
     const navigate = useNavigate()
@@ -22,9 +22,11 @@ export const usePusherListeners = (
             setOpponentName(opponentName)
             navigate("/play")
         })
+
         gameManager.events.addOnGetCardsListener((cards: Card[]) => {
             setCards(cards)
         })
+
         gameManager.events.addOnMyTurnStartListener(() => {
             console.log("my turn start")
             const possibleActions = gameManager.getPossibleActions()
@@ -37,14 +39,37 @@ export const usePusherListeners = (
             setActions(newActions)
             setIsMyTurn(true)
         })
+
         gameManager.events.addOnMyTurnEndListener(() => {
             console.log("my turn end")
             setIsMyTurn(false)
             setActions([])
         })
 
-        gameManager.events.addOnMyPlayCardListener((card: Card) => {
+        gameManager.events.addOnCardPlayedListener((iCalled: boolean, card: Card) => {
+            if (!iCalled) {
+                console.log("opponent played card: " + card.number + " " + card.suit)
+                return;
+            }
+            console.log("i played card: " + card.number + " " + card.suit)
             setCards((cards: Card[]) => cards.filter((c) => c.number !== card.number || c.suit !== card.suit))
+        })
+
+        gameManager.events.addOnGameEndListener((IWon: boolean) => {
+            console.log("game end")
+            console.log(IWon)
+            setIsMyTurn(false)
+            setActions([])
+            setOpponentName("")
+            setCards([])
+            setTimeout(() => {
+                navigate("/") // TODO - replace with a modal saying who won
+            }, 2000)
+        })
+        gameManager.events.addOnPointsUpdateListener((myPoints: number, opponentPoints: number) => {
+            console.log("points update")
+            console.log(myPoints)
+            console.log(opponentPoints)
         })
     }, [])
 }
